Moved `examples` task from ENB config
